combine input and output module to inout
