--- src/pycropml/input.py
@@ -1,5 +1,0 @@
-
-import inputoutput as hh
-class Input(hh.InputOutput):
-    """ Input """
-
