--- src/pycropml/output.py
@@ -1,5 +1,0 @@
-
-import inputoutput as hh
-class Output(hh.InputOutput):
-    """ Output
-    """
